+++ actions/metadata/feature_stats.py
@@ -4,7 +4,7 @@ from copy import deepcopy
 import gin
 import numpy as np
 
-from actions.utils import get_parse_filter_text
+from actions.util_functions import get_parse_filter_text
 
 
 def compute_stats(df, labels, f, conversation):
