+++ actions/metadata/labels.py
@@ -1,7 +1,7 @@
 """Show data labels"""
 import gin
 
-from actions.utils import get_parse_filter_text
+from actions.util_functions import get_parse_filter_text
 
 
 @gin.configurable
