+++ actions/prediction/predict.py
@@ -1,7 +1,7 @@
 """Prediction operation."""
 import numpy as np
 
-from actions.utils import gen_parse_op_text, get_parse_filter_text
+from actions.util_functions import gen_parse_op_text, get_parse_filter_text
 
 
 def handle_input(parse_text):
