+++ templates/index.html
@@ -36,8 +36,8 @@
 
         <form class="ttm-inputarea">
             <select id="input_source" class="form-select" style="width: 150px; text-align:center; font-family: Calibri;">
-                <option style="text-align:center; " selected value="0">input</option>
-                <option style="text-align:center;" value="1">customed input</option>
+                <option style="text-align:center; " selected value="0">Input</option>
+                <option style="text-align:center;" value="1">Custom input</option>
             </select>
 
             <input style='margin-left: 10px'type="text" class="ttm-input" id="textInput"
